requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
asyncer==0.0.11
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
boto3==1.37.17
botocore==1.37.17
cachetools==6.2.2
certifi==2025.1.31
cffi==2.0.0
chainlit==2.9.3
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
code2flow==2.5.1
colorama==0.4.6
CounterFit==0.1.4.dev9
cryptography==46.0.3
cuid==0.4
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
email_validator==2.2.0
eventlet==0.33.3
fastapi==0.124.4
filelock==3.20.0
filetype==1.2.0
Flask==2.1.2
Flask-SocketIO==5.2.0
frozenlist==1.5.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.1.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.14.0
hf-xet==1.2.0
httpcore==1.0.7
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface_hub==1.2.3
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.48
langchain-openai==0.3.10
langchain-text-splitters==0.3.7
langsmith==0.3.18
Lazify==0.4.0
literalai==0.1.201
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.24.0
multidict==6.2.0
mypy-extensions==1.0.0
narwhals==2.13.0
nest-asyncio==1.6.0
numpy==2.2.4
openai==1.68.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-agno==0.49.8
opentelemetry-instrumentation-alephalpha==0.49.8
opentelemetry-instrumentation-anthropic==0.49.8
opentelemetry-instrumentation-bedrock==0.49.8
opentelemetry-instrumentation-chromadb==0.49.8
opentelemetry-instrumentation-cohere==0.49.8
opentelemetry-instrumentation-crewai==0.49.8
opentelemetry-instrumentation-google-generativeai==0.49.8
opentelemetry-instrumentation-groq==0.49.8
opentelemetry-instrumentation-haystack==0.49.8
opentelemetry-instrumentation-lancedb==0.49.8
opentelemetry-instrumentation-langchain==0.49.8
opentelemetry-instrumentation-llamaindex==0.49.8
opentelemetry-instrumentation-logging==0.60b1
opentelemetry-instrumentation-marqo==0.49.8
opentelemetry-instrumentation-mcp==0.49.8
opentelemetry-instrumentation-milvus==0.49.8
opentelemetry-instrumentation-mistralai==0.49.8
opentelemetry-instrumentation-ollama==0.49.8
opentelemetry-instrumentation-openai==0.49.8
opentelemetry-instrumentation-openai-agents==0.49.8
opentelemetry-instrumentation-pinecone==0.49.8
opentelemetry-instrumentation-qdrant==0.49.8
opentelemetry-instrumentation-redis==0.60b1
opentelemetry-instrumentation-replicate==0.49.8
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-sagemaker==0.49.8
opentelemetry-instrumentation-sqlalchemy==0.60b1
opentelemetry-instrumentation-threading==0.60b1
opentelemetry-instrumentation-together==0.49.8
opentelemetry-instrumentation-transformers==0.49.8
opentelemetry-instrumentation-urllib3==0.60b1
opentelemetry-instrumentation-vertexai==0.49.8
opentelemetry-instrumentation-watsonx==0.49.8
opentelemetry-instrumentation-weaviate==0.49.8
opentelemetry-instrumentation-writer==0.49.8
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.60b1
orjson==3.10.16
packaging==24.2
pandas==2.3.3
pillow==12.0.0
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==6.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.15.0
python-telegram-bot==21.10
pytz==2025.2
pywin32==311
PyYAML==6.0.2
RapidFuzz==3.14.3
referencing==0.37.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.11.4
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.39
sse-starlette==3.0.3
starlette==0.50.0
streamlit==1.52.1
streamlit-lottie==0.0.5
syncer==2.0.3
tenacity==9.0.0
thefuzz==0.22.1
tiktoken==0.9.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
tornado==6.5.3
tqdm==4.67.1
traceloop-sdk==0.49.8
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
uuid6==2024.7.10
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==0.24.0
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
yarl==1.18.3
zipp==3.23.0
zstandard==0.23.0
